Validate Records Function Update
Function checks a Json format datastructure.  It updates the properties
and values based on the parameters passed into the function.  Added
updates for if a property is passed in and doesnt exist.  If the tracks
property doesnt it creates an empty array to hold the tracks then adds
the track value to it.  More to come...