Function Inventory Update #2
Added function that loops through the arrays and sets variables to the
different keys and values required for later use.  Added loop that
checks if keys from arr1 and arr2 match.  If they do it updates their
values and pushes them to the final array.  Added function that pushes
remaining arr2 elems to the final array.  Mostly working.