Pairwise Function Update
Created a function that loops through the array elements and for each i loops through all array elements >= to the current i.  It then checks the addition of the element with the current i against the input argument.  If it equals the argument it adds the element indices to an answer variable.